backend/langgraph/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spell"
version = "0.1.0"
description = "Spell project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain-community>=0.3.19",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.5",
    "readabilipy>=0.3.0",
    "python-dotenv>=1.0.1",
    "socksio>=1.0.0",
    "markdownify>=1.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "python-multipart>=0.0.9",
    "sse-starlette>=1.6.5",
    "pandas>=2.2.3",
    "numpy>=2.2.3",
    "yfinance>=0.2.54",
    "langchain-deepseek>=0.1.2",
    "langchain-google-genai>=2.0.8",
    "google-genai>=0.3.0",
    "langchain-google-community>=2.0.10",
    "langgraph-checkpoint-postgres>=3.0.2",
    "psycopg[binary,pool]>=3.2.0",
    "google-cloud-storage>=2.19.0",
    "pydantic-settings>=2.7.0",
    # PPTXテンプレート機能
    "python-pptx>=1.0.0",
    "lxml>=5.0.0",
    "pdf2image>=1.17.0",
    "Pillow>=10.0.0",
]


[project.optional-dependencies]
dev = [
    "black>=24.2.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/build/
'''
